+++ src/interfaces/plugins/index.ts
@@ -11,6 +11,7 @@ export * from './optionsHandler/optionsHandler.interface';
 export * from './optionsHandler/simpleOptionsHandler.interface';
 export * from './popup/popup.interface';
 export * from './popup/simplePopup.interface';
+export * from './positioner/popoverPositioner.interface';
 export * from './positioner/positioner.interface';
 export * from './readonlyState/readonlyState.interface';
 export * from './valueHandler/dynamicValueHandler.interface';