+++ src/interfaces/plugins/positioner/popoverPositioner.interface.ts
@@ -1,0 +1,14 @@
+import {PositionerOptions} from './positioner.interface';
+
+/**
+ * Options for popover positioner plugin
+ */
+export interface PopoverPositionerOptions extends PositionerOptions
+{
+    //######################### properties #########################
+
+    /**
+     * Indication whether popover property should be set to 'auto' or 'manual'
+     */
+    popoverAuto: boolean;
+}