+++ src/plugins/positioner/index.ts
@@ -1,1 +1,3 @@
 export * from './commonPositioner/commonPositioner.component';
+export * from './noPositioner/noPositioner.component';
+export * from './popoverPositioner/popoverPositioner.component';